[go-gen] Move request/response types from DAO in Match service, and increase consistency

+++ match/comm/handler.go
@@ -9,7 +9,7 @@ import (
 
 // Comm provides the interface adopted by Handler, allowing for mocking
 type Comm interface {
-	CheckUser(userID int) (bool, error)
+	CheckUser(userID int64) (bool, error)
 }
 
 // Handler maintains the list of services and their associated hostnames
@@ -23,7 +23,7 @@ func Init(config *util.Config) *Handler {
 }
 
 // CheckUser makes a request to the target service to check if a user ID exists
-func (comm *Handler) CheckUser(userID int) (bool, error) {
+func (comm *Handler) CheckUser(userID int64) (bool, error) {
 	hostname, ok := comm.Services["user"]
 	if !ok {
 		return false, fmt.Errorf("service %s's hostname not in config file", "user")
